Addressing passive zap endpoint thread starvation
When a passive endpoint was assigned to a thread that has an active
endpoint with high activity, the routine of the passive endpoint (e.g.
accepting a connection) could be starved. The routines of the passive
endpoints are quite light-weighted. So, this patch assigns the passive
endpoints into its own thread to avoid the starvation.